extract UseLazyEmbed from BpModal
